修改字段类型

+++ pay-external-api-server/src/main/java/com/tfjt/pay/external/unionpay/api/service/impl/LoanApiServiceImpl.java
@@ -154,7 +154,7 @@ public class LoanApiServiceImpl implements LoanApiService {
     }
 
     @Override
-    public Result<DepositRespDTO> deposit(Integer amount, String orderNo) {
+    public Result<DepositRespDTO> deposit(Long amount, String orderNo) {
         return loanUserBizService.deposit(amount, orderNo);
     }
