+++ pay-external-service/src/main/java/com/tfjt/pay/external/unionpay/dto/req/ProductInfoReqDTO.java
@@ -21,7 +21,7 @@ public class ProductInfoReqDTO {
     /**
      * 订单金额
      */
-    private Integer orderAmount;
+    private Long orderAmount;
 
     /**
      * 商品名称
